Start of rgba refactor
